+++ CHANGELOG.md
@@ -1,18 +1,29 @@
+### v7.0.4
+
+#### Fix
+
+- fix swap providers [#5367](https://github.com/MyEtherWallet/MyEtherWallet/pull/5367)
+
+#### Feat
+
+- mew wallet banner [#5340](https://github.com/MyEtherWallet/MyEtherWallet/pull/5340)
+
 ### v7.0.3
 
-### Fix
+#### Fix
 
 - fix metamask error on cancelling send and fix bug when disconnecting through metamask [#5337](https://github.com/MyEtherWallet/MyEtherWallet/pull/5337)
 - handle crap token send error [#5330](https://github.com/MyEtherWallet/MyEtherWallet/pull/5330)
 - Add filter low fiat balance tokens [#5333](https://github.com/MyEtherWallet/MyEtherWallet/pull/5333)
 - fix font loading [#5356](https://github.com/MyEtherWallet/MyEtherWallet/pull/5356)
 
-# Feat
+#### Feat
 
 - Add blinking info for approval on ledger or trezor [#5331](https://github.com/MyEtherWallet/MyEtherWallet/pull/5331)
 - store selected pair in swap and trade, and set it once user navigates back [#5335](https://github.com/MyEtherWallet/MyEtherWallet/pull/5335)
 - store chosen network on bridge [#5343](https://github.com/MyEtherWallet/MyEtherWallet/pull/5343)
 - fix error when watchlist tokens are cleared [#5336](https://github.com/MyEtherWallet/MyEtherWallet/pull/5336)
+- Add Amplitude tracking events for stock market, crypto market, trade, and swap sorting interactions [#5365](https://github.com/MyEtherWallet/MyEtherWallet/pull/5365)
 
 ### v7.0.2